+++ frontend/src/Components/Modal/ModalHeader.js
@@ -1,3 +1,4 @@
+import classNames from 'classnames';
 import PropTypes from 'prop-types';
 import React, { Component } from 'react';
 import styles from './ModalHeader.css';
@@ -9,13 +10,14 @@ class ModalHeader extends Component {
 
   render() {
     const {
+      className,
       children,
       ...otherProps
     } = this.props;
 
     return (
       <div
-        className={styles.modalHeader}
+        className={classNames(styles.modalHeader, className)}
         {...otherProps}
       >
         {children}
@@ -26,6 +28,7 @@ class ModalHeader extends Component {
 }
 
 ModalHeader.propTypes = {
+  className: PropTypes.string,
   children: PropTypes.node
 };